==> 02_scripts/04_Treina_KMeans_e_salva_modelo.ipynb <==
# 04_Treina_KMeans_e_salva_modelo

## 01_import_bibliotecas

import pandas as pd
from sklearn.cluster import KMeans
import joblib
import os
from sklearn.preprocessing import StandardScaler

## 02_Carga_de_dados

input = 'https://github.com/Kinrider/tech_challenge_5/raw/refs/heads/main/01_fontes/arquivos_decision/fontes_tratadas/03_input_clusterizacao.parquet'

df = pd.read_parquet(input)

## 03_Treina_modelo

# === 2. Tratar valores ausentes e treinar o modelo KMeans com k=4 ===
colunas_com_nan = df.columns[df.isnull().any()].tolist()
for col in colunas_com_nan:
    df[f"tem_{col}"] = df[col].notnull().astype(int)
    df[col] = df[col].fillna(-9999)
    
kmeans = KMeans(n_clusters=4, random_state=42, n_init=10)
df['cluster'] = kmeans.fit_predict(df)

# === 3. Salvar colunas utilizadas no input do modelo ===
colunas_usadas = df.drop(columns=['cluster']).columns.tolist()

## 04_Salvando_modelo

# === 4. Definir diretório de saída ===
caminho_saida = r"C:\Users\pedro\Documents\Área de Trabalho\tech_challenge_5\03_modelos"
os.makedirs(caminho_saida, exist_ok=True)

# === 5. Salvar arquivos ===
joblib.dump(kmeans, os.path.join(caminho_saida, "modelo_kmeans.joblib"))
joblib.dump(colunas_usadas, os.path.join(caminho_saida, "colunas_usadas.joblib"))
df.to_parquet(os.path.join(caminho_saida, "04_clusterizados.parquet"), index=False)


# Variáveis contínuas usadas no modelo
variaveis_continuas = ['remuneracao_zscore', 'tempo_experiencia_anos', 'quantidade_experiencias']

# Treinar o scaler
scaler = StandardScaler()
scaler.fit(df[variaveis_continuas])

# Salvar o scaler
joblib.dump(scaler,os.path.join(caminho_saida,"scaler.joblib"))

print("✅ Modelo treinado e arquivos salvos em:")
print(caminho_saida)
