==> 02_scripts/05_aux_remuneracao.ipynb <==
import pandas as pd

input_cluster = 'https://github.com/Kinrider/tech_challenge_5/raw/refs/heads/main/01_fontes/arquivos_decision/fontes_tratadas/02_candidatos_tratados.parquet'

df = pd.read_parquet(input_cluster)

df.head()

df_filtrado = df[["remuneracao_mensal_brl"]]

Caminho = "C:\\Users\\pedro\\Documents\\Área de Trabalho\\tech_challenge_5\\01_fontes\\arquivos_decision\\fontes_tratadas\\"

df_filtrado.to_parquet(Caminho + "remuneracao_mensal_only.parquet", index=False)
